Upgrade to Lightstep Terraform provider version 1.84.1
What does this PR do?
---------------------

- Upgrades all modules to https://registry.terraform.io/providers/lightstep/lightstep/1.84.1
- Updates auto-gen files

For what purpose?
---------------------

The dashboards and modules in this repo have been pinned to provider version 1.79.0. However, some newer dashboards in this repo have been built with features that were introduced in more recent Lightstep TF provider versions. All the versions in this repo should be consistent, especially so that `terraform init` works. v1.84.1 is the latest Lightstep TF provider version